chore(userEvent): extract duplicate functions
